move report command to corresponding group folder
